requirements.txt
Flask==2.0.1
google-cloud-vision==2.3.1
google-cloud-firestore==2.1.1